TSification of codediff.js
