+++ difflib.js
@@ -34,22 +34,22 @@ var difflib = {
 	defaultJunkFunction: function (c) {
 		return __whitespace.hasOwnProperty(c);
 	},
-	
+
 	stripLinebreaks: function (str) { return str.replace(/^[\n\r]*|[\n\r]*$/g, ""); },
-	
+
 	stringAsLines: function (str) {
 		var lfpos = str.indexOf("\n");
 		var crpos = str.indexOf("\r");
 		var linebreak = ((lfpos > -1 && crpos > -1) || crpos < 0) ? "\n" : "\r";
-		
+
 		var lines = str.split(linebreak);
 		for (var i = 0; i < lines.length; i++) {
 			lines[i] = difflib.stripLinebreaks(lines[i]);
 		}
-		
+
 		return lines;
 	},
-	
+
 	// iteration-based reduce implementation
 	__reduce: function (func, list, initial) {
 		if (initial != null) {
@@ -61,14 +61,14 @@ var difflib = {
 		} else {
 			return null;
 		}
-		
+
 		for (; idx < list.length; idx++) {
 			value = func(value, list[idx]);
 		}
-		
+
 		return value;
 	},
-	
+
 	// comparison function for sorting lists of numeric tuples
 	__ntuplecomp: function (a, b) {
 		var mlen = Math.max(a.length, b.length);
@@ -76,45 +76,45 @@ var difflib = {
 			if (a[i] < b[i]) return -1;
 			if (a[i] > b[i]) return 1;
 		}
-		
+
 		return a.length == b.length ? 0 : (a.length < b.length ? -1 : 1);
 	},
-	
+
 	__calculate_ratio: function (matches, length) {
 		return length ? 2.0 * matches / length : 1.0;
 	},
-	
+
 	// returns a function that returns true if a key passed to the returned function
 	// is in the dict (js object) provided to this function; replaces being able to
 	// carry around dict.has_key in python...
 	__isindict: function (dict) {
 		return function (key) { return dict.hasOwnProperty(key); };
 	},
-	
+
 	// replacement for python's dict.get function -- need easy default values
 	__dictget: function (dict, key, defaultValue) {
 		return dict.hasOwnProperty(key) ? dict[key] : defaultValue;
-	},	
-	
+	},
+
 	SequenceMatcher: function (a, b, isjunk) {
 		this.set_seqs = function (a, b) {
 			this.set_seq1(a);
 			this.set_seq2(b);
 		}
-		
+
 		this.set_seq1 = function (a) {
 			if (a == this.a) return;
 			this.a = a;
 			this.matching_blocks = this.opcodes = null;
 		}
-		
+
 		this.set_seq2 = function (b) {
 			if (b == this.b) return;
 			this.b = b;
 			this.matching_blocks = this.opcodes = this.fullbcount = null;
 			this.__chain_b();
 		}
-		
+
 		this.__chain_b = function () {
 			var b = this.b;
 			var n = b.length;
@@ -134,13 +134,13 @@ var difflib = {
 					b2j[elt] = [i];
 				}
 			}
-	
+
 			for (var elt in populardict) {
 				if (populardict.hasOwnProperty(elt)) {
 					delete b2j[elt];
 				}
 			}
-			
+
 			var isjunk = this.isjunk;
 			var junkdict = {};
 			if (isjunk) {
@@ -157,11 +157,11 @@ var difflib = {
 					}
 				}
 			}
-	
+
 			this.isbjunk = difflib.__isindict(junkdict);
 			this.isbpopular = difflib.__isindict(populardict);
 		}
-		
+
 		this.find_longest_match = function (alo, ahi, blo, bhi) {
 			var a = this.a;
 			var b = this.b;
@@ -172,7 +172,7 @@ var difflib = {
 			var bestsize = 0;
 			var j = null;
 			var k;
-	
+
 			var j2len = {};
 			var nothing = [];
 			for (var i = alo; i < ahi; i++) {
@@ -193,38 +193,38 @@ var difflib = {
 				}
 				j2len = newj2len;
 			}
-	
+
 			while (besti > alo && bestj > blo && !isbjunk(b[bestj - 1]) && a[besti - 1] == b[bestj - 1]) {
 				besti--;
 				bestj--;
 				bestsize++;
 			}
-				
+
 			while (besti + bestsize < ahi && bestj + bestsize < bhi &&
 					!isbjunk(b[bestj + bestsize]) &&
 					a[besti + bestsize] == b[bestj + bestsize]) {
 				bestsize++;
 			}
-	
+
 			while (besti > alo && bestj > blo && isbjunk(b[bestj - 1]) && a[besti - 1] == b[bestj - 1]) {
 				besti--;
 				bestj--;
 				bestsize++;
 			}
-			
+
 			while (besti + bestsize < ahi && bestj + bestsize < bhi && isbjunk(b[bestj + bestsize]) &&
 					a[besti + bestsize] == b[bestj + bestsize]) {
 				bestsize++;
 			}
-	
+
 			return [besti, bestj, bestsize];
 		}
-		
+
 		this.get_matching_blocks = function () {
 			if (this.matching_blocks != null) return this.matching_blocks;
 			var la = this.a.length;
 			var lb = this.b.length;
-	
+
 			var queue = [[0, la, 0, lb]];
 			var matching_blocks = [];
 			var alo, ahi, blo, bhi, qi, i, j, k, x;
@@ -238,7 +238,7 @@ var difflib = {
 				i = x[0];
 				j = x[1];
 				k = x[2];
-	
+
 				if (k) {
 					matching_blocks.push(x);
 					if (alo < i && blo < j)
@@ -247,9 +247,9 @@ var difflib = {
 						queue.push([i + k, ahi, j + k, bhi]);
 				}
 			}
-			
+
 			matching_blocks.sort(difflib.__ntuplecomp);
-	
+
 			var i1 = 0, j1 = 0, k1 = 0, block = 0;
 			var i2, j2, k2;
 			var non_adjacent = [];
@@ -269,14 +269,14 @@ var difflib = {
 					}
 				}
 			}
-			
+
 			if (k1) non_adjacent.push([i1, j1, k1]);
-	
+
 			non_adjacent.push([la, lb, 0]);
 			this.matching_blocks = non_adjacent;
 			return this.matching_blocks;
 		}
-		
+
 		this.get_opcodes = function () {
 			if (this.opcodes != null) return this.opcodes;
 			var i = 0;
@@ -302,14 +302,14 @@ var difflib = {
 					if (tag) answer.push([tag, i, ai, j, bj]);
 					i = ai + size;
 					j = bj + size;
-					
+
 					if (size) answer.push(['equal', ai, i, bj, j]);
 				}
 			}
-			
+
 			return answer;
 		}
-		
+
 		// this is a generator function in the python lib, which of course is not supported in javascript
 		// the reimplementation builds up the grouped opcodes into a list in their entirety and returns that.
 		this.get_grouped_opcodes = function (n) {
@@ -335,7 +335,7 @@ var difflib = {
 				j2 = code[4];
 				codes[codes.length - 1] = [tag, i1, Math.min(i2, i1 + n), j1, Math.min(j2, j1 + n)];
 			}
-	
+
 			var nn = n + n;
 			var group = [];
 			var groups = [];
@@ -354,23 +354,23 @@ var difflib = {
 						i1 = Math.max(i1, i2-n);
 						j1 = Math.max(j1, j2-n);
 					}
-					
+
 					group.push([tag, i1, i2, j1, j2]);
 				}
 			}
-			
+
 			if (group && !(group.length == 1 && group[0][0] == 'equal')) groups.push(group)
-			
+
 			return groups;
 		}
-		
+
 		this.ratio = function () {
 			matches = difflib.__reduce(
 							function (sum, triple) { return sum + triple[triple.length - 1]; },
 							this.get_matching_blocks(), 0);
 			return difflib.__calculate_ratio(matches, this.a.length + this.b.length);
 		}
-		
+
 		this.quick_ratio = function () {
 			var fullbcount, elt;
 			if (this.fullbcount == null) {
@@ -381,7 +381,7 @@ var difflib = {
 				}
 			}
 			fullbcount = this.fullbcount;
-	
+
 			var avail = {};
 			var availhas = difflib.__isindict(avail);
 			var matches = numb = 0;
@@ -395,16 +395,16 @@ var difflib = {
 				avail[elt] = numb - 1;
 				if (numb > 0) matches++;
 			}
-			
+
 			return difflib.__calculate_ratio(matches, this.a.length + this.b.length);
 		}
-		
+
 		this.real_quick_ratio = function () {
 			var la = this.a.length;
 			var lb = this.b.length;
 			return _calculate_ratio(Math.min(la, lb), la + lb);
 		}
-		
+
 		this.isjunk = isjunk ? isjunk : difflib.defaultJunkFunction;
 		this.a = this.b = null;
 		this.set_seqs(a, b);
