+++ package.json
@@ -7,10 +7,12 @@
     "url": "https://github.com/danvk/codediff.js.git"
   },
   "dependencies": {
-    "jquery": ">=2.0.0",
-    "highlight.js": "~8.0.0"
+    "highlight.js": "~8.0.0",
+    "jquery": ">=2.0.0"
   },
   "devDependencies": {
+    "@types/highlightjs": "^9.12.2",
+    "@types/jquery": "^3.5.14",
     "assert": "*",
     "grunt": "~0.4.0",
     "grunt-cli": "0.1.6",
@@ -18,7 +20,8 @@
     "grunt-qunit-cov": "~0.3.6",
     "grunt-qunit-junit": "~0.1.0",
     "phantomjs": "~1.9.7",
-    "qunitjs": "^1.14.0"
+    "qunitjs": "^1.14.0",
+    "typescript": "^4.8.2"
   },
   "scripts": {
     "test": "grunt travis"
